Microsoft/PTC/News/ppt_EN_to_KR/requirements.txt
python-pptx==0.6.23
openai==1.75.0
python-dotenv==1.1.0
tqdm==4.67.1
